Makefile now tries to fix bad precision NNs
Generated neural networks that are outside of the bounds of what the
accelerator can handle will now automatically be converted to networks
that we can handle. This should fix an issue that can periodically crop
up when generating randomized NNs for testing.

The Makefile will now use a single seed to generate all NNs.