Adds ability to set ASID and NNID in fann-batch
This includes the ability to set both the ASID and NNID when running
fann-batch. Both the ASID and NNID table are filled out with either
empty configuration regions or garbage values around the one valid NNID.

The regression test 03-fann-batch now uses this with a randomized ASID
and NNID in range zero to nine.