Cleanup Makefiles, scripts, and regression tests
`fann-float-to-fixed` and `write-fann-config-for-accelerator` now no
longer spew debug information about the neural networks. This can be
re-enabled with a '-v' flag to either of them. To enable this, both of
these were migrated to the more recent getopt style that I've been
using.

The Makefiles do not enable this verbosity. Other approaches to decrease
the verbosity of the output have also been added, e.g., the Makefiles
now don't print bash `if` blocks to stdout.

The top-level Makefile now builds the "rv" target by default, instead of
"all" which would try to do testing of X-Files/Dana in Chisel. This
Makefile now determines the number of parallel jobs to run using the
`max-processors.sh` script.