Add learning parameter, new config, TTable assert
This adds a new "learning" parameter that will eventually change whether
or not the resulting X-Files/Dana system supports learning. This
includes a new configuration, XFilesDanaNoLearningCPPConfig which builds
such a variant. Currently, nothing is changed between this and a normal
configuration architecturally, however, the Transaction Table of a "no
learning" configuration will fire off an assertion if it sees a learning
request come in.