Use "Learn" PE->Cache Interface, preemptively
The PE Table now uses the learning variant of the PE->Cache interface
even though this is the same interface as the non-learning one. This
may change in the future as the non-learning variant can technically
use smaller field sizes due to fewer PE->Cache request types actually
being used when a non-learning variant is built.