Kludge to fix xfiles-supervisor for broken fseek
With the current version of the RISC-V toolchain, fseek to SEEK_END
does not seem to work. This replaces the one fseek in
xfiles-supervisor.c with a while loop fread.

This also adds error checking to both xfiles-supervisor and fann-batch
which will cause these to die if xfiles-supervisor sees a file size of
0.

More visibility as to what is going on related to this is included in
the ANTW Chisel code.