Update rvcon for specific FPGA, fix rvreboot help
This makes it possible to request a specific FPGA with a command line
argument to rvcon. This is useful if you need to, for some reason,
connect to a specific FPGA (e.g., to update something in its u-boot),
but don't want to grab all the locks to do so.

Minor update to the rvreboot help text.