Change binary format, saved PREVIOUS layer size
The previous layer size is needed when doing error
backpropagation---the address of the previous layer's inputs must be
known in order to compute the slope updates. This enables the
possibility for that (no HDL changes, yet) by overwriting the unused
next layer sizing with the previous layer sizing. The previous layer
size for the input layer is defined to be the number of inputs to the
NN.