Use correct exit condition for incremental
When running incremental learning, the output layer is never read during
the weight update step. The original exit condition checks to see if the
user process has read all the outputs in the "current layer" which would
be the penultimate layer and not the last layer. So, the learning
variant now stores the number of nodes in the output layer during
feedforward computation and uses this to determine if the exit criteria
has been met.