Fix feedforward regression, cleanup assignments
Non-learning feedforward computation does not have knowledge of
inLastEarly, so the number of nodes in the last layer is blindly set
upon receipt of a layer information.

This also cleans up some assignments in the Transaction Table by using
the val `nicl` instead of the explicit Cache response data vec.