Add methods for dumping bundle (table) state
This adds two methods to DanaBundle (`printElements` and `printAll`)
which will print the names and values of all entries in a Bundle. These
are used by an `info` method in DanaModule which dumps information about
a specific table (a vector of DanaBundle). An example of how to use this
is shown in the Cache.