Respond when no ASID, add X-FILES/DANA ID
When the X-FILES arbiter receives a new write request on an invalid
ASID, the arbiter squashes the request from getting to the Transaction
Table and responds with a value of -1. A basic C program is included
that demonstrates this functionality.

This adds a new X-FILES/DANA ID function that will respond with some
information about the X-FILES/DANA configuration. The request is
squashed in the same fashion as with an invalid new write request.

The request types now use enumerated types defined in both the X-FILES
software libraries and in the abstract XFilesModule class.