Register File, Cache addressing fixes
The Register File was overprovisioned in terms of SRAM depth. There
exists one SRAM per Transaction Table entry, but each SRAM was being
created as if it was being sized as if it was the only SRAM.

The global weight pointer was being set during the first response
pipeline stage when the global weight pointer was not yet ready. This
resulted in a random initial location for slope storage. An assertion
that will fire on writes to addresses outside of the SRAM is now in
place to catch this.