Fix Register File addressing and assertion
A fix introduced in the Register File addressing was incorrect. The
Register File is composed of separate memories and indexing should not
depend on a TID computed offset.

This corrects an incorrect assertion that was causing failures for large
networks related to checking Register File memory access bounds. The
Register File deals with element indices, but the assertion was checking
as if these were block indices.