Simplify public interface for dana-memory-tool
This gives dana-memory-tool a private learning attribute which defines
whether or not the memory layout printed will correspond to a learning
transaction. The public API now uses a single `printMemory` method which
will print either a feedforward or learning layout based on the
initialization of the Memory object.

Add help text showing usage with `column`.