Move helper functions of fann-xfiles to a header
This creates a new file that contains the helper functions used by
fann-xfiles.c:
 * src/test/rv/fann-xfiles.h

This also adds the top level directory ("." from the perspective of the
top level Makefile) to the include path when building test programs.
This allows files to follow the Google style guide for C/C++ where the
path to a header is fully specified.