Remove dependency on all FANN builds for tools
This explicitly sets the dependency for building the tools as only
relying on the host machine version of FANN. This previously was set
as requiring every version of FANN, i.e., host/x86 as well as RISC-V
newlib and linux-gnu. This resulted in any build rule which required a
tool as needing to have both a newlib and linux-gnu install of the
RISC-V tools. Specifically, this would cause Travis to barf when
running fann-xfiles regression tests.