Add explicit XFilesBackend class
This then makes DANA extend XFilesBackend. The interface between X-FILES
hardware and DANA then uses the XFilesBackendInterface.

The XFilesBackendInterface is currently DANA-specific in that it
includes multiple RoCC command lines for DANA's Transaction Table and
ASID--NNID Table Walker. This should be reduced to a single
RoCCInterface with internal routing handled on DANA.