Single D$ port to backend, refactor ANTW
This enables a single data cache (`mem`) port connection to the backend.
Routing and arbitration between multiple cores is then handled by
changes to the XFilesArbiter. This includes outbound request routing,
queuing of responses, and arbitration amongst response queues with valid
data. The single port interface is then used by the ASID--NNID Table
Walker (ANTW).

This also includes a refactor of the ANTW to use a more "Chisel-like"
structure. The state logic is not placed in a switch statement, but is
handled be sequentially structured `when` blocks.

This increases the size of the ANTW Config ROB from 16 to 32 to address
issues related to overwriting entries in the Config ROB when the ANTW is
allowed to generate requests full bore (#5), i.e., whenever the Data
Cache is ready. This may be a fix for current Rocket configuration
parameters, but needs a generic fix that either stops requests that
could overwrite from hitting the data cache or dealing with them when
they come back, e.g., with Config ROB associativity.