Fix ANTW NNID offset bug
This corrects a problem where the NNID was not being used to determine
the NN configuration index. The NNID was then effectively always zero.

This adds a check that the found NN configuration size is nonzero. This
will cause the ANTW to fail faster in the event of it reading a garbage
NNID Table entry.