Update Proxy Kernel patch with ANTP/ASID returns
This modifies the X-FILES syscalls of the Proxy Kernel so that they will
return the old value of the ASID--NNID Table Pointer (ANTP) or ASID when
a new value is written. A special value will be returned by the hardware
in the event that an invalid ASID or TID is swapped out.