Use single RoCCInterface between X-FILES/Backend
This changes the interface of the XFilesBackendInterface to use a single
RoCCInterface instead of separate DANA-specific interfaces (a partial
RoCCInterface for both the ASID--NNID Table Walker and DANA's
Transaction Table). Core indices are passed using a new CoreIdx Bundle
with one CoreIdx Bundle for the register file, `cmd`/`resp` interface
and another for the L1 Dcache `mem` interface.