Delay Cache notifyFlag to allow SRAM to update
Using an SRAM that takes more than one cycle to writeback data inside
of the Configuration Cache can cause the Configuration Cache to send
garbage (old data) to the Transaction Table. This only happens,
however, for memories that take more than one cycle to writeback,
e.g., the memory used for the learning variant of the Configuration
Cache.

This makes the base Configuration Cache class abstract with the
definition of `fIsDoneFetching` undefined. The non-learning variant
uses the old `fIsDoneFetching` function, while the learning variant
adds an additional cycle of delay to ensure that the Cache memory has
valid data.