XF TTable helper funcs, responses from XF TTable
This changes all the DANA Transaction Table printfs to follows the
convention of "DANA TTable: ...".

The TID response format has been changed to remove a dependency on the
X-FILES Transaction Table needing to know DANA's elementWidth parameter.
The TID is now, simply, placed in the LSBs of the xLen response.

This adds some helper functions to the X-FILES Transaction Table for
generating responses to the microprocessor and for pulling data out of
the correct locations of RoCC data fields.

The data passing convention for the response to a new write request has
been changed.

This migrates the new request response (that returns the TID to the
Rocket core) from DANA's Transaction Table to the X-FILES Transaction
Table.