Put rs2 newRequest data in XF TTable Input Queue
The backend needs some initial information about what the request is.
This could be accomplished by an additional data line going from the XF
TTable directly to the backend. However, since the input/output queue
infrastructure already exists, the contents of the rs2 register are now
place directly into the input queue on a new request.