Single bath OK, fix exit code, output sourcing
The major changes here are the proper handling of exit conditions and
when batches are deemed finished and when the whole batch is done. This
does not appear to be working for feedforward computations on a
learning-enabled DANA.

The value returned by `write_data` of src/main/c/xfiles-user.c was
incorrectly returning the entire output instead of just the 3-bit exit
code. The response from X-FILES hardware returns the TID which, for
non-zero TIDs, will cause the exit code to be non-zero. This now returns
only the 3-bit exit code.

The sourcing of expected output data for a learning transaction was
being pulled from rs1 instead of rs2. This is now corrected.