Connect new MStatus line of updated Rocc
The RoCCInterface was updated to pass all of the CSRs to the RoCC
backend through an MStatus bundle. This gets X-FILES/DANA to build, but
without verifying the functionality of X-FILES/DANA.

Modules which use the new MStatus just operate on an or reduction of the
MStatus "prv" field, i.e., a reduction that yields whether we're in a
non-user mode.

The RoCCInterface status (class MStatus) was previously unconnected.
This is now fixed, kludgily, by setting the privilege bits of the status
field to that of the requesting core if there is supervisor command
forwarding or to zero otherwise.