Makefile updates for top level, regression tests
This updates the Makefiles to fix some breakages in the toolflow for the
regression tests.

The RISC-V binary-specific build steps, that cover the generation of a
RISC-V executable, have been moved to a separated Makefile fragment in
tools/common/Makefrag-rv.

The regression build flow now works and should only build what it needs.
Additionally, the C++ emulator can now be built from inside the lowest
level Makefiles (e.g., tests/tests/XX-YYY/Makefile). This should not be
needed as the Berkeley stamps-based Makefile will build the needed C++
emulator.