Cleanup imports, genericize backend instantiation
Cosmetic cleanup of all imports to pull in only what is needed. Cosmetic
cleanup of imports that pull in one class/object/thing to not use curly
braces.

This genericizes the instantiation of the backend where the backend is
passed as an XFilesBackendParameters class containing two fields. This
approach is modeled after the way that RoCC backends are already
instantiated. The two fields are:
  * generator -- a function that creates an XFilesBackend from a set of
    parameters.
  * info -- an (xLen - 16) bit value that will be printed as the LSBs
    when a "get info" request is received by the X-FILES arbiter. This
    is intended to allow a user process to get some information about
    the backend. This will default to zero if unspecified.

The current "get info" function, xfiles_dana_id, has been updated to
match the new info format.