Update rvreboot for PCIe boards, fix "fpga" bug
This adds support for boards which are not connected to the APC, but
are directly installed in a PCIe slot for the rvreboot script. Note,
that this just fails if you try to reboot a board which doesn't have
plug info.

This also corrects a small bug where you could specif a FPGA as "fpga"
and it would match the first entry in the info.txt file. The whole
name of the FPGA should be provided.