Refactor libxfiles into user and supervisor libs
This provides separately built libraries for user and supervisor
portions of the X-FILES libraries:
  * libxfiles-user.a
  * libxfiles-supervisor.a

The intent here is to provide a distinct supervisory library that can be
pulled in by the Proxy Kernel or Linux that only defines supervisory
functions (small assembly functions for accessing registers on the
accelerator). The earlier previous supervisory library included Proxy
Kernel-specific system calls for doing things which the kernel is
supposed to do and was littered with printfs (which the kernel isn't
going to like). This moves all that code into the user library with the
intent that this will eventually be removed.

The Proxy Kernel patch is updated to use functions defined in and link
against the new X-FILES supervisor library.