Add DebugUnit with reg read/write, debug-test.c
This adds a per-core Debug Unit inside the X-FILES Arbiter. This is
intended to facilitate testing of the Linux kernel migration (#3) by
allowing for explicit reads/writes of addresses to verify that we have
the virtual/physical model correct.

This unit can currently only be used as a register request/response
unit.