Add PTW support to DebugUnit
The X-FILES DebugUnit now supports address translation using the PTW
port, i.e., a request can be generated that will perform address
translation of a given address. Support for this new function has been
added to the X-FILES user library and included in the debug-test
program.

Note: currently the address translation succeeds, however, the returned
address is the same as the virtual address when tested using the Proxy
Kernel. This may be expected or is possibly indicative of something
incorrect on our end.