Change usr/sup req numbering, fix funct fragility
This makes the user and supervisor request function codes (the "funct"
of the XCustom instruction) sequentially numbered with supervisor
requests reserved for the two LSBs:
  * supervisor [0, 3]
  * user [4, 2^7-1]

Note: that there is no optimization for efficient comparison, here!

This change caught a lingering function code bug where DANA's
Transaction Table was reading off a specific bit of the function code to
determine if this was a write of the last data element.