Fix riscv-pk patch to build pk without xfiles libs
The old riscv-pk patch required an existing version of the X-FILES/DANA
supervisor libraries to already be installed. This introduced a circular
dependency that conflicted with how the Rocket Chip regression tests
build things:
  1) ./riscv-tools/build.sh (depended on XF/D libs)
  2) build xfiles-dana

This is now corrected in that the necessary files from the X-FILES/DANA
libraries (specifically, xfiles-supervisor.[c,h]) are now symlinked into
the Proxy Kernel machine directory and built along with the machine
code.