Align NN configuration on L2 cache line boundary
This changes the data structure for each NNID table entry with raw and
L2 cache line aligned pointers. The latter is used for the actual NN
configuration and is what the hardware ASID--NNID Table Walker uses.

This is a breaking changed that requires simultaneously upgrading the
hardware and the X-FILES library.