Remove explicit INPUT/OUTPUT in PE (bug fix)
This removes all the explicit declarations of INPUT or OUTPUT in
ProcessingElement.scala. This fixes a problem where Decoupled/Valid
would not properly flip these signals correctly. As a result, inputs
would show up as outputs in the PE Table and last-connect semantics
would overwrite these. In effect, the PEs were left largely unconnected
to the PE Table.