Use new reqAf function for PE->AF communication
This adds and uses a new method, reqAf, in the Processing Element (PE)
that uses a non-deterministic, two-state write req followed by read resp
to communicate with the Activation Function (AF). This mirrors the way
that the PE communicates with the PE Table.

Minor cleanup to both the PE and AF.