XF/D Configs -> src/main/scala/Configs.scala
Remove the top-level "configs" directory and just put the X-FILES/DANA
Configs file in the src/main/scala/ directory. Since this is already in
`package rocketchip` it can live anywhere and does not need to be
symlinked into the explicit src/ directory of rocket chip. The unneeded
`install-symlinks` script is, consequently removed.