Bug fix in use of libxfiles Virt->Phys, cleanup
Using the X-FILES' debug unit to do virtual to physical address
translation can fail in the way that it was used in libxfiles.
Specifically, doing address translation of an address in a malloc'd
region will cause a page fault if the address that you're trying to read
goes across a page boundary. To avoid this issue (as we have no
intention of handling a page fault from the X-FILES debug unit), the
neural network configuration is copied into the aligned region before
the address translation. This will ensure that the memcpy will cause all
the page faults and our hardware access to the PTW will succeed.

This additionally cleans up libxfiles to use local pointers to keep
dereferences shorter.