Add standalone XFiles configuration
This adds a standalone top module containing just X-FILES and a
generator to emit FIRRTL for this standalone module. This, with some
future changes, will enable testing of X-FILES/DANA outside of a
rocket-chip environment.

Note: in order for this to work, this requires dummy configuration
parameters to be defined that are normally defined by a Rocket Chip
configuration.