Get Exuberant Ctags working
This adds support for a `make tags` target which will use Exuberant
Ctags to build an Emacs-supported "TAGS" file in the top level
directory. Modifications for Vim could be made.

Note, this relies on having an installed version of Exuberant Ctags (not
Emacs ctags/etags or whatever the package manager gives you).
Additionally, you need to have Scala language definitions available in
~/.ctags.

This will build TAGS for Scala and C sources and will build the tags for
Chisel3 and Rocket Chip to provide all definitions that you could
possibly need.

Running `make clean` will now remove the built TAGS file if it exists.