For regressions, install tools in rocket-chip
Currently these tools are built in xfiles-dana/tests/install. This is
okay for standard regression testing, but on Travis this directory is
cached. With this directory cached the submodule clone of xfiles-dana
fails as Travis sets up the xfiles-dana/tests/install directory before
doing the clone. A clone into a non-empty directory will die fast.