Refactor build flow directory structure
This is in preparation for building a standalone version of the
libraries for use on the host system (needed to do standalone simulation
work involving the ASID--NNID table). This changes the directory
structure for the builds and their ./usr symlinks to use the explicit
target name for what is being built. This matches the target is a
Makefile parameter, e.g., the following will build everything using the
Linux GNU toolchain:
  make TARGET=riscv64-unknown-linux-gnu

The build structure is now:
  build
  |--> fann
  |    |--> host
  |    |--> riscv64-unknown-elf
  |    \--> riscv64-unknown-linux-gnu
  |--> riscv64-unknown-elf
  \--> riscv64-unknown-linux-gnu

The ./usr structure is now:
  usr
  |--> riscv64-unknown-elf
  |    \--> lib
  \--> riscv64-unknown-linux-gnu
       \--> lib