Update build for host build without virtual memory
The host system is only expected to work with a simulated standalone
xfiles-dana. The host system is responsible for populating an ASID--NNID
Table and loading this in a dummy BRAM for DANA to consume. This means
that a subset of the full-blown xfiles-dana libraries need to be built
only containing functions for manipulating the ASID-NNID Table. This
requires some ifdefs to remove any uses sage of `debug_virt_to_phys` in
the code. This makes these modifications and updates the build system in
the tools directory to create a host version of this library and use it
for a new program, `generate-ant`. This is a WIP program to populate an
ASID--NNID Table with whatever the user wants.