Use PREFIX (ROCC-dir) for tests
This updates the Makefiles to use a PREFIX instead of just the ROCC as
the initial portion of the string that describes a test. This can be
best described via enumerating the resulting changes. The tests are
now named as follows when built:
  * build/smoke/xfiles-dana-p-debug -> build/smoke/xfiles-dana-smoke-p-debug
  * build/smoke/xfiles-dana-p-id -> build/smoke/xfiles-dana-smoke-p-id
  * build/nets/xfiles-dana-p-xorSigmoidSymmetric ->
      build/nets/xfiles-dana-nets-p-xorSigmoidSymmetric

The fact that multiple tests in different directories shared the same
PREFIX made things difficult for the Makefile fragment emitted for the
emulator to properly tell these apart.

This changes are achieved with the use of a new PREFIX variable that
superseded the old ROCC variable.