New Python tool for generating asm .data ANTs
This adds a new tool (overloaded on naming) that can be used to
generate an ASID--NNID Table (ANT) from a set of ASID,NNID tuples.
This mirrors the origin generate-ant.c, except that this generates the
ANT directly and does not use X-FILES/DANA libraries on the host
machine to do it.

This avoids all issues of alignment encountered previously and results
in a more compact ANT representation.