SMP test spot check, fix printfs, perfect-chisel
This fixes a long-outstanding issue where the notify mask from
dana.Cache to dana.TransactionTable was not used. This meant that SMP
would not work for transactions with the same [ASID, NNID] tuple where
the later transaction showed up while the first transaction was still
fetching data from L2.

With this RTL modification, the new `xorSigmoidSymmetric-smp` test is
passing.

This includes more cleanups to the printfs and the inclusion of the
`perfect-chisel` submodule.