Remove ID test from regression suite
Temporarily disable the ID test from the regression test suite. This
test is a difficult one as it's checking that the XF/D ID string
matches what is expected. The ID string needs to be statically known
when building the tests based on the configuration we're using. This
is fine for Travis (as it is building separate tests for each entry in
the build matrix), but is bothersome for running regressions locally
as usually only one instance of the tests build directory exists.