Finish xorSigmoidSymmetric and -smp tests
Using the comparison added to the SMP fann_run function in submodule
'libs', this finishes the xorSigmoidSymmetric tests.

The xorSigmoidSymmetric-smp test has been rolled into it's non-smp
variant via the use of one preprocessor macro.

Note: these tests are non-optimal and should not be used for any kind
of benchmarking.