Add abstract CSR File, remove ASID, ANTP pointer
Add and use X-FILES and Backend (DANA) CSRs. This consolidates the
ASID and ANTP/size registers inside of this unified CSR File. This,
consequently, completely removes the AsidUnit and migrates all ASID
functionality into an X-FILES CSR.

This modifies the Transaction Table (on the X-FILES side) to only
allow a number of simultaneous transactions as those defined in a CSR.