Fixes for DANA CSRs, connect CSR status lines
This adds DANA CSRs, properly sets their bit widths (see note below)
and includes a generic `status` Bundle that is propagated back to DANA
for, as intended but not used, controlling backend functionality at
run time.

Note: the use of `lazy val` is required with the way that we're mixing
the CSR of X-FILES and the backend together. We need the registers to
be made available to functions in the parent X-FILES CSR file.
However, this results in the parameters used by these modules being
out of scope. As a result, the parameters need to be fully qualified
and can't use something that only exists in a Backend module parent,
but not an X-FILES parent.