[WIP] Add partial cache writeback, weight decay CSR
This adds support for writing data back into L2. The data looks fine
going out over the AUTL interface, but has not been verified that it
does, in fact, exist in memory.

This adds a weight decay CSR that is initialized to 1.0 (i.e., the
weight is multiplied by 1.0 for everyone weight update cycle).

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@ibm.com>