Fixes for incremental learning spot-check
This modifies Dana's Transaction Table state signal `isFirst` such
that this is updated only when a layer comes back as valid. This
contrasts with its earlier definition as updating when the last PE in
a layer is allocated. This was required for checks when the layer
comes back as valid to handle incremental learning.

This was not tested for batch learning and may result in regressions
there.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@ibm.com>