Revert to fpga-zynq aligning repo, tests cleanup
This reverts all of xfiles-dana to match a version of rocket-chip
that can be interfaced with fpga-zynq/master and, assumedly, the
Hurricane branch of rocket-chip. Reversions include:
  * Switching from `config` to `cde`
  * ANTW increments the addr on a put block for TileLink(1)

The unnecessary dependency on `config` in perfect-chisel is removed.

The tests are refactored to point directly at the
environment (riscv/riscv-env) directory included in a top-level
rocket-chip as opposed to using their own. This removes the need for
keeping these consistent.

The tests are refactored to put common Makefile code in a top-level
Makefrag and to follow standard Makefile naming conventions.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@ibm.com>