Fix bad Cache -> ANTW queue extraction
This fixes a bug where using a configuration with a single
Configuration Cache entry would cause the ANTW never to budge from its
initial state. This, broadly, appears to be a bug in how a
single-entry Queue is described. This avoids that problem by, more
accurately, just using the dequeue valid signal to determine if a
cache request is available as opposed to the count signal.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@ibm.com>