Fix multi-output network bug
This fixes a bug in the determination of when a transaction at the
arbiter level is deemed to be complete (and is evicted from the
arbiter's transaction table). A simultaneous cache response from the
backend to an output queue and a read of that queue would cause the
transaction to be deemed complete and evicted. The initiating RISC-V
program would then be stuck in a spinlock and would always fail with
the transaction being reported as invalid.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@ibm.com>