Fix output queue bug
This fixes a bug in the output queue where the incorrect queue index
is used to check for queue readiness (on the almost full line). This
resulted in deadlocks for SMP workloads.

An assertion is now provided that will check for this deadlock
condition (an empty output queue for a transaction waiting for output
data).

This includes a minor refactor of the Arbiter Transaction Table where
printfs and assertions are mixed in as traits by a companion object.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@ibm.com>