Use structs to describe DANA binary configuration
This modifies `write-fann-config-for-accelerator` to use structs to
describe the [global, layer, info] regions a binary configuration for
DANA. This dramatically simplifies the understanding of what is going
on here and makes explicit documentation (in
doc/binary-encodings-data-structures.md) much less important.

This has been tested as generating the exact same configurations for
all available networks.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@ibm.com>