Define and use largest possible width
This adds a new method in DanaParameters that returns the largest
possible width from the three non-weight regions of the neural network
configuration. This is then used to set the required bandwidth between
the Cache and the Transaction Table. This should remove fragility on
this definition if we want to play around with further modifications
to the ABI in terms of sizing the configuration regions differently.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@ibm.com>